moving stuff out of headers
