+++ Tools/interface/Normalizer.h
@@ -28,35 +28,7 @@ public:
      * \brief Normalizes the tree given total weight of events and 
      * cross-section
      */
-    void NormalizeTree(TTree *t, double totalEvts, double xsec) {
-        float inWeight=1, outWeight=1;
-        double inWeightD=1;
-        t->SetBranchStatus("*",0);
-        turnOnBranches(t,inWeightName.Data());
-        TBranch *b = t->Branch(outWeightName.Data(),&outWeight,TString::Format("%s/F",outWeightName.Data()));
-        if (inWeightName!="") {
-           if (isFloat)    
-             t->SetBranchAddress(inWeightName.Data(),&inWeight);
-           else
-            t->SetBranchAddress(inWeightName.Data(),&inWeightD);
-        } else {
-            inWeight = 1;
-            inWeightD = 1;
-        }
-        unsigned int nEntries = t->GetEntries();
-        unsigned int iE=0;
-        ProgressReporter pr("Normalizer::NormalizeTree",&iE,&nEntries,10);
-        for (iE=0; iE!=nEntries; ++iE) {
-            pr.Report();
-            t->GetEntry(iE);
-            if (isFloat)
-                outWeight = xsec*inWeight/totalEvts;
-            else
-                outWeight = xsec*inWeightD/totalEvts;
-            b->Fill();
-        }
-        t->SetBranchStatus("*",1);
-    }
+    void NormalizeTree(TTree *t, double totalEvts, double xsec);
 
     /**
      * \param fpath path to input file 
@@ -64,18 +36,7 @@ public:
      * \brief Reads an input file and picks up the tree to normalize
      * as well as a histogram containing the weight of events
      */
-    void NormalizeTree(TString fpath, double xsec) {
-        TFile *fIn = TFile::Open(fpath,"UPDATE");
-        TTree *t = (TTree*)fIn->Get(treeName.Data());
-        TH1F *h = (TH1F*)fIn->Get(histName.Data());
-        if (t==NULL || h==NULL) {
-            PError("Normalizer::NormalizeTree",TString::Format("Could not normalize %s because tree=%p and hist=%p\n",fpath.Data(),t,h));
-            return; 
-        }
-        NormalizeTree(t,h->Integral(),xsec);
-        fIn->WriteTObject(t);
-        fIn->Close();
-    }
+    void NormalizeTree(TString fpath, double xsec);
 
     TString inWeightName = "mcWeight"; /**< name of input branch */
     TString outWeightName = "normalizedWeight"; /**< name of output branch */
