+++ Tools/interface/TreeTools.h
@@ -16,14 +16,7 @@
  * \brief Figures out any substrings of s that are branch dependencies
  * and sets the status of those branches to 1
  */
-inline void turnOnBranches(TTree*t,TString s) {
-	std::vector<TString> deps = getDependencies(s);
-	for (auto d : deps) {
-		TBranch *b = (TBranch*)t->GetListOfBranches()->FindObject(d);
-		if (b)
-			t->SetBranchStatus(d,1);
-	}
-}
+void turnOnBranches(TTree*t,TString s);
 
 /**
  * \param t input chain
@@ -32,14 +25,7 @@ inline void turnOnBranches(TTree*t,TString s) {
  * and sets the status of those branches to 1. This is meant for
  * TChains becacuse SetBranchStatus behaves differently
  */
-inline void turnOnBranchesChain(TChain*t,TString s) {
-	std::vector<TString> deps = getDependencies(s);
-	for (auto d : deps) {
-		TBranch *b = (TBranch*)t->GetListOfBranches()->FindObject(d);
-		if (b)
-			t->SetBranchStatus(d,1);
-	}
-}
+void turnOnBranchesChain(TChain*t,TString s);
 
 //////////////////////////////////////////////////////////////////////////////////
